# Topk refactor