--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -1,3 +1,8 @@
+ # Part of this file was contributed by NVIDIA under license:
+ #   Copyright (C) 2020 NVIDIA Corporation
+ #   SPDX-License-Identifier: MIT
+
+add_definitions(-DCUB_IGNORE_DEPRECATED_CPP_DIALECT=1)
 add_subdirectory(3rd_party)
 
 include_directories(.)