--- a/src/tensors/gpu/topk.cu
+++ b/src/tensors/gpu/topk.cu
@@ -1,3 +1,8 @@
+ /* All or part of this file was contributed by NVIDIA under license:
+ *   Copyright (C) 2020 NVIDIA Corporation
+ *   SPDX-License-Identifier: MIT
+ */
+
 #include "common/definitions.h"
 #include "tensors/tensor_operators.h"
 #include "tensors/gpu/cuda_helpers.h"