--- a/src/translator/nth_element.cu
+++ b/src/translator/nth_element.cu
@@ -3,6 +3,11 @@
  *   SPDX-License-Identifier: MIT
  */
 
+ /* All or part of this file was contributed by NVIDIA under license:
+ *   Copyright (C) 2020 NVIDIA Corporation
+ *   SPDX-License-Identifier: MIT
+ */
+
 #include <iostream>
 
 #include "translator/nth_element.h"