@rhenry-nv 2020-12-19
I think I have addressed all of your feedback @frankseide.

I added detailed comments about the functions and made the stylistic changes you suggested. Let me if there's anything else that you think needs to be changed.